+++ src/index.ts
@@ -1,8 +1,7 @@
-import * as openApi from './open-api';
 import * as validation from './validation';
 import * as fastify from './fastify';
 import * as pagination from './pagination';
 import * as cardano from './cardano';
 
 export { MetricsCollector } from './metrics-collector';
-export { openApi, validation, fastify, pagination, cardano };
+export { validation, fastify, pagination, cardano };