+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+
+- `getCIPstandard`, `getOnchainMetadata`, `getOnchainMetadataVersion` functions
+
+### Removed 
+
+- BREAKING CHANGE: `getSchemaForEndpoint` is now part of the `@blockrost/openapi` package
+
 ## [1.1.0] - 2022-11-02
 
 ### Added