+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0] - 2022-11-27
 
-### Removed 
+### Removed
 
 - BREAKING CHANGE: `getSchemaForEndpoint` is now part of the `@blockrost/openapi` package
 - `@blockrost/openapi` dependency